refactor(history): extract exercise logs chart
